Fix #95, traceback off by two lines
This squash-merged commit addresses #95, line numbers in Repy tracebacks are off by 2.

The fix involves creating a new module, `encoding_header.py`, which solely exists to declare the encoding header that is appended to source files in order to avoid possible `encode`/`decode` issues (see SeattleTestbed/ATTIC#982). This encoding header -- and the number of lines it contains -- is used in multiple other modules, and having this single central declaration avoids circular import headaches.

Usages of the header include subtracting its length from line numbers that appear in tracebacks (which is what #95 reports); we also append this very declaration as the actual encoding header for user code.

Other contributions in this commit include
* unit tests that check traceback line numbers by causing errors in specific lines,
* fixes for #115 so that better error messages are produced on  specific errors in `safe.py`, and
* general coding style cleanups.

Thanks go out to @JustinCappos for an initial take on the fix, @awwad and @lukpueh for additional work on the implementation, and  @vladimir-v-diaz for helpful review comments and testing.