Fix #132, deadly errors from within `VirtualNamespace`
This commit fixes an issue in `namespace.py`'s handling of the
wrapped `evaluate` method of `virtual_namespace.VirtualNamespace`.
`namespace.py` previously assumed that all RepyV2 API calls will
only ever raise `RepyException`s (or subclasses thereof). Since
PR #127 made virtual namespaces wrapped objects, the earlier
assumption is invalid: A virtual namespace can raise plain Python
exceptions (like `NameError`s) as well!

We therefore add a `VirtualNamespace`-specific clause for handling
Python exceptions raised by `NamespaceAPIFunctionWrapper.wrapped_function`s:
If a Python exception is seen, and it originates from an object of
type `VirtualNamespace`, then we reraise the exception so that
user code can see and handle it. For other origins, consider the
Python exception an indication of a programming error in the sandbox
API definitions, and terminate the sandbox.

This commit also adds a unit test that raises a `NameError` inside
of a `VirtualNamespace` to demonstrate the problem and its fix.