Fix #123, file close concurrency unit test
This commit refactors "ut_repyv2api_filecloseconcurrency.r2py", adds
a new restrictions file for it, and adapts the amount of bytes written
and read so that it is more likely for the assumptions of this test
(q.v.) to hold.

The test checks that `writeat` and `readat` of a RepyV2 file still
succeed even if the file is `close`d in a concurrent thread. Yet,
actually providing a concurrent thread is not as easy as it sounds.
In particular, the test assumes that starting a thread is "quick"
which is true for most but not all Seattle platforms, OpenWrt boxes
being a notable exception.

Therefore, the test now writes and reads larger amounts of data
so that it is more likely that the thread started by the test
actually runs while read/write operations are still in progress,
as desired.